Add support for Target Retirement Age, Employee pension contribution and Employer contribution.
Removed the simple RetirementCalculator.

+++ TaxCalculator/ExternalInterface/IIncomeTaxCalculator.cs
@@ -2,6 +2,6 @@
 {
     public interface IIncomeTaxCalculator
     {
-        ITaxResult TaxFor(int payeSalary);
+        ITaxResult TaxFor(decimal payeSalary);
     }
 }
